requirements.txt
# Core web framework
fastapi
uvicorn[standard]
python-multipart
websockets
python-dotenv
pydantic

# Neural TTS engine deps
torch
numpy
soundfile
scipy
transformers
huggingface_hub
loguru
misaki[en,ja,zh]>=0.9.4

# PDF engine
PyMuPDF

# STT engine
openai-whisper

# Audio utils
pydub
webrtcvad-wheels
